+++ src/main/java/com/shadowinlife/app/SQLModelFactory/BuildingUpgrade.java
@@ -1,5 +1,4 @@
 package com.shadowinlife.app.SQLModelFactory;
-
 import java.io.*;
 import java.sql.*;
 
@@ -60,11 +59,11 @@ public class BuildingUpgrade extends BaseBean implements Serializable {
         return vUin;
     }
 
-    private Integer iRoleId;
-    public void setiRoleId(Integer iRoleId){
+    private Long iRoleId;
+    public void setiRoleId(Long iRoleId){
         this.iRoleId=iRoleId;
     }
-   public Integer getiRoleId() {
+   public Long getiRoleId() {
         return iRoleId;
     }
 
@@ -76,43 +75,43 @@ public class BuildingUpgrade extends BaseBean implements Serializable {
         return vRoleName;
     }
 
-    private Integer iRoleJob;
-    public void setiRoleJob(Integer iRoleJob){
+    private Long iRoleJob;
+    public void setiRoleJob(Long iRoleJob){
         this.iRoleJob=iRoleJob;
     }
-   public Integer getiRoleJob() {
+   public Long getiRoleJob() {
         return iRoleJob;
     }
 
-    private Integer iRoleGender;
-    public void setiRoleGender(Integer iRoleGender){
+    private Long iRoleGender;
+    public void setiRoleGender(Long iRoleGender){
         this.iRoleGender=iRoleGender;
     }
-   public Integer getiRoleGender() {
+   public Long getiRoleGender() {
         return iRoleGender;
     }
 
-    private Integer iRoleLevel;
-    public void setiRoleLevel(Integer iRoleLevel){
+    private Long iRoleLevel;
+    public void setiRoleLevel(Long iRoleLevel){
         this.iRoleLevel=iRoleLevel;
     }
-   public Integer getiRoleLevel() {
+   public Long getiRoleLevel() {
         return iRoleLevel;
     }
 
-    private Integer iRoleVipLevel;
-    public void setiRoleVipLevel(Integer iRoleVipLevel){
+    private Long iRoleVipLevel;
+    public void setiRoleVipLevel(Long iRoleVipLevel){
         this.iRoleVipLevel=iRoleVipLevel;
     }
-   public Integer getiRoleVipLevel() {
+   public Long getiRoleVipLevel() {
         return iRoleVipLevel;
     }
 
-    private Integer iRoleReputationLevel;
-    public void setiRoleReputationLevel(Integer iRoleReputationLevel){
+    private Long iRoleReputationLevel;
+    public void setiRoleReputationLevel(Long iRoleReputationLevel){
         this.iRoleReputationLevel=iRoleReputationLevel;
     }
-   public Integer getiRoleReputationLevel() {
+   public Long getiRoleReputationLevel() {
         return iRoleReputationLevel;
     }
 
@@ -132,27 +131,27 @@ public class BuildingUpgrade extends BaseBean implements Serializable {
         return vRoleElse2;
     }
 
-    private Integer iRoleExperience;
-    public void setiRoleExperience(Integer iRoleExperience){
+    private Long iRoleExperience;
+    public void setiRoleExperience(Long iRoleExperience){
         this.iRoleExperience=iRoleExperience;
     }
-   public Integer getiRoleExperience() {
+   public Long getiRoleExperience() {
         return iRoleExperience;
     }
 
-    private Integer iRoleSword;
-    public void setiRoleSword(Integer iRoleSword){
+    private Long iRoleSword;
+    public void setiRoleSword(Long iRoleSword){
         this.iRoleSword=iRoleSword;
     }
-   public Integer getiRoleSword() {
+   public Long getiRoleSword() {
         return iRoleSword;
     }
 
-    private Integer iPartyId;
-    public void setiPartyId(Integer iPartyId){
+    private Long iPartyId;
+    public void setiPartyId(Long iPartyId){
         this.iPartyId=iPartyId;
     }
-   public Integer getiPartyId() {
+   public Long getiPartyId() {
         return iPartyId;
     }
 
@@ -164,27 +163,27 @@ public class BuildingUpgrade extends BaseBean implements Serializable {
         return vPartyName;
     }
 
-    private Integer iPartyLevel;
-    public void setiPartyLevel(Integer iPartyLevel){
+    private Long iPartyLevel;
+    public void setiPartyLevel(Long iPartyLevel){
         this.iPartyLevel=iPartyLevel;
     }
-   public Integer getiPartyLevel() {
+   public Long getiPartyLevel() {
         return iPartyLevel;
     }
 
-    private Integer iBuildingParty;
-    public void setiBuildingParty(Integer iBuildingParty){
+    private Long iBuildingParty;
+    public void setiBuildingParty(Long iBuildingParty){
         this.iBuildingParty=iBuildingParty;
     }
-   public Integer getiBuildingParty() {
+   public Long getiBuildingParty() {
         return iBuildingParty;
     }
 
-    private Integer iUpgrateBulidingId;
-    public void setiUpgrateBulidingId(Integer iUpgrateBulidingId){
+    private Long iUpgrateBulidingId;
+    public void setiUpgrateBulidingId(Long iUpgrateBulidingId){
         this.iUpgrateBulidingId=iUpgrateBulidingId;
     }
-   public Integer getiUpgrateBulidingId() {
+   public Long getiUpgrateBulidingId() {
         return iUpgrateBulidingId;
     }
 
@@ -196,19 +195,19 @@ public class BuildingUpgrade extends BaseBean implements Serializable {
         return vUpgrateBulidingName;
     }
 
-    private Integer iBulidingLevel;
-    public void setiBulidingLevel(Integer iBulidingLevel){
+    private Long iBulidingLevel;
+    public void setiBulidingLevel(Long iBulidingLevel){
         this.iBulidingLevel=iBulidingLevel;
     }
-   public Integer getiBulidingLevel() {
+   public Long getiBulidingLevel() {
         return iBulidingLevel;
     }
 
-    private Integer iUpgrateCostParty;
-    public void setiUpgrateCostParty(Integer iUpgrateCostParty){
+    private Long iUpgrateCostParty;
+    public void setiUpgrateCostParty(Long iUpgrateCostParty){
         this.iUpgrateCostParty=iUpgrateCostParty;
     }
-   public Integer getiUpgrateCostParty() {
+   public Long getiUpgrateCostParty() {
         return iUpgrateCostParty;
     }
 
@@ -228,19 +227,19 @@ public class BuildingUpgrade extends BaseBean implements Serializable {
         return jPartyLeftMoney;
     }
 
-    private Integer iPartyLeftPolity;
-    public void setiPartyLeftPolity(Integer iPartyLeftPolity){
+    private Long iPartyLeftPolity;
+    public void setiPartyLeftPolity(Long iPartyLeftPolity){
         this.iPartyLeftPolity=iPartyLeftPolity;
     }
-   public Integer getiPartyLeftPolity() {
+   public Long getiPartyLeftPolity() {
         return iPartyLeftPolity;
     }
 
-    private Integer iCostTime;
-    public void setiCostTime(Integer iCostTime){
+    private Long iCostTime;
+    public void setiCostTime(Long iCostTime){
         this.iCostTime=iCostTime;
     }
-   public Integer getiCostTime() {
+   public Long getiCostTime() {
         return iCostTime;
     }
 
@@ -264,29 +263,29 @@ public class BuildingUpgrade extends BaseBean implements Serializable {
         this.iEventId = Long.valueOf(iEventId);
         this.vVersionId = String.valueOf(vVersionId);
         this.vUin = String.valueOf(vUin);
-        this.iRoleId = Integer.valueOf(iRoleId);
+        this.iRoleId = Long.valueOf(iRoleId);
         this.vRoleName = String.valueOf(vRoleName);
-        this.iRoleJob = Integer.valueOf(iRoleJob);
-        this.iRoleGender = Integer.valueOf(iRoleGender);
-        this.iRoleLevel = Integer.valueOf(iRoleLevel);
-        this.iRoleVipLevel = Integer.valueOf(iRoleVipLevel);
-        this.iRoleReputationLevel = Integer.valueOf(iRoleReputationLevel);
+        this.iRoleJob = Long.valueOf(iRoleJob);
+        this.iRoleGender = Long.valueOf(iRoleGender);
+        this.iRoleLevel = Long.valueOf(iRoleLevel);
+        this.iRoleVipLevel = Long.valueOf(iRoleVipLevel);
+        this.iRoleReputationLevel = Long.valueOf(iRoleReputationLevel);
         this.vRoleElse1 = String.valueOf(vRoleElse1);
         this.vRoleElse2 = String.valueOf(vRoleElse2);
-        this.iRoleExperience = Integer.valueOf(iRoleExperience);
-        this.iRoleSword = Integer.valueOf(iRoleSword);
-        this.iPartyId = Integer.valueOf(iPartyId);
+        this.iRoleExperience = Long.valueOf(iRoleExperience);
+        this.iRoleSword = Long.valueOf(iRoleSword);
+        this.iPartyId = Long.valueOf(iPartyId);
         this.vPartyName = String.valueOf(vPartyName);
-        this.iPartyLevel = Integer.valueOf(iPartyLevel);
-        this.iBuildingParty = Integer.valueOf(iBuildingParty);
-        this.iUpgrateBulidingId = Integer.valueOf(iUpgrateBulidingId);
+        this.iPartyLevel = Long.valueOf(iPartyLevel);
+        this.iBuildingParty = Long.valueOf(iBuildingParty);
+        this.iUpgrateBulidingId = Long.valueOf(iUpgrateBulidingId);
         this.vUpgrateBulidingName = String.valueOf(vUpgrateBulidingName);
-        this.iBulidingLevel = Integer.valueOf(iBulidingLevel);
-        this.iUpgrateCostParty = Integer.valueOf(iUpgrateCostParty);
+        this.iBulidingLevel = Long.valueOf(iBulidingLevel);
+        this.iUpgrateCostParty = Long.valueOf(iUpgrateCostParty);
         this.jUpgrateCostMoney = String.valueOf(jUpgrateCostMoney);
         this.jPartyLeftMoney = String.valueOf(jPartyLeftMoney);
-        this.iPartyLeftPolity = Integer.valueOf(iPartyLeftPolity);
-        this.iCostTime = Integer.valueOf(iCostTime);
+        this.iPartyLeftPolity = Long.valueOf(iPartyLeftPolity);
+        this.iCostTime = Long.valueOf(iCostTime);
         this.dtEventEndTime = Timestamp.valueOf(dtEventEndTime);
     }
 }
